% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_poos_recursively.R
\name{forecast_poos_recursively.BVARPANEL}
\alias{forecast_poos_recursively.BVARPANEL}
\title{Bayesian recursive pseudo-out-of-sample forecasting}
\usage{
\method{forecast_poos_recursively}{BVARPANEL}(model_spec, poos_spec, show_progress = TRUE)
}
\arguments{
\item{model_spec}{an object of class \code{BVARPANEL} generated using the 
\code{specify_bvarPANEL} function and containing the Bayesian Panel VAR model 
specification.}

\item{poos_spec}{an object of class \code{POOSForecastSetup} containing
specification of the recursive pseudo-out-of-sample forecasting exercise 
using expanding window samples.}

\item{show_progress}{a logical value, if \code{TRUE} the estimation progress bar is visible}
}
\value{
An object of class \code{ForecastsPOOS} containing the outcome of Bayesian
 recursive pseudo-out-of-sample forecasting exercise  using expanding window 
 samples containing elements
}
\description{
Performs the recursive pseudo-out-of-sample forecasting exercise
using expanding window samples.
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
