% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_poos_recursively.R
\name{forecast_poos_recursively}
\alias{forecast_poos_recursively}
\title{Bayesian recursive pseudo-out-of-sample forecasting}
\usage{
forecast_poos_recursively(model_spec, poos_spec, show_progress = TRUE)
}
\arguments{
\item{model_spec}{an object generated using one of the \code{specify_*} 
functions containing model specification.}

\item{poos_spec}{an object of class \code{POOSForecastSetup} containing
specification of the recursive pseudo-out-of-sample forecasting exercise 
using expanding window samples.}

\item{show_progress}{a logical value, if \code{TRUE} the estimation progress bar is visible}
}
\value{
An object of class \code{ForecastsPOOS} containing the outcome of Bayesian
 recursive pseudo-out-of-sample forecasting exercise  using expanding window 
 samples containing elements:
\describe{
 \item{forecasts}{an array with the draws from predictive density}
 \item{forecast_mean}{an array with the mean of the predictive densities}
 \item{forecast_cov}{an array with the covariances of the predictive densities}
 \item{estimation_data}{an array with data used for estimation for each forecasts}
 \item{evaluation_data}{an array with matched data for forecasting performance evaluation}
}
}
\description{
Performs the recursive pseudo-out-of-sample forecasting exercise
using expanding window samples.
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
