% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_bvargrouppanel.R
\name{specify_posterior_bvarGroupPANEL}
\alias{specify_posterior_bvarGroupPANEL}
\title{R6 Class Representing PosteriorBVARGROUPPANEL}
\description{
The class PosteriorBVARGROUPPANEL contains posterior output and the specification 
including the last MCMC draw for the Bayesian Panel VAR model with country grouping. 
Note that due to the thinning of the MCMC output the starting value in element 
\code{last_draw} might not be equal to the last draw provided in 
element \code{posterior}.
}
\examples{
spec = specify_bvarGroupPANEL$new(
   data = ilo_dynamic_panel[1:5],
   G = 2
)
#posterior       = estimate(specification, 5)
#class(posterior)

}
\seealso{
\code{\link{specify_bvarGroupPANEL}}
}
\section{Super class}{
\code{PosteriorBVARPANEL} -> \code{PosteriorBVARGROUPPANEL}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{last_draw}}{an object of class BVARGROUPPANEL with the last draw of the 
current MCMC run as the starting value to be passed to the continuation 
of the MCMC estimation using \code{estimate()}.}

    \item{\code{posterior}}{a list containing Bayesian estimation output.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-PosteriorBVARGROUPPANEL-initialize}{\code{PosteriorBVARGROUPPANEL$new()}}
    \item \href{#method-PosteriorBVARGROUPPANEL-clone}{\code{PosteriorBVARGROUPPANEL$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><code>PosteriorBVARPANEL$get_last_draw()</code></li>
  <li><code>PosteriorBVARPANEL$get_posterior()</code></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PosteriorBVARGROUPPANEL-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-PosteriorBVARGROUPPANEL-initialize}{}}}
\subsection{\code{PosteriorBVARGROUPPANEL$new()}}{
  Create a new posterior output PosteriorBVARGROUPPANEL.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PosteriorBVARGROUPPANEL$new(specification, posterior)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{specification}}{an object of class BVARGROUPPANEL with the last 
draw of the current MCMC run as the starting value.}
      \item{\code{posterior}}{a list containing Bayesian estimation output.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A posterior output PosteriorBVARGROUPPANEL.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PosteriorBVARGROUPPANEL-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PosteriorBVARGROUPPANEL-clone}{}}}
\subsection{\code{PosteriorBVARGROUPPANEL$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PosteriorBVARGROUPPANEL$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
