% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_bvargrouppriorpanel.R
\name{specify_posterior_bvarGroupPriorPANEL}
\alias{specify_posterior_bvarGroupPriorPANEL}
\title{R6 Class Representing PosteriorBVARGROUPPRIORPANEL}
\description{
The class PosteriorBVARGROUPPRIORPANEL contains posterior output and the specification 
including the last MCMC draw for the Bayesian Panel VAR model with country grouping. 
Note that due to the thinning of the MCMC output the starting value in element 
\code{last_draw} might not be equal to the last draw provided in 
element \code{posterior}.
}
\examples{
spec = specify_bvarGroupPriorPANEL$new(
   data = ilo_dynamic_panel[1:5],
   G = 2
)
#posterior       = estimate(specification, 5)
#class(posterior)

}
\seealso{
\code{\link{specify_bvarGroupPriorPANEL}}
}
\section{Super classes}{
\code{bpvars::PosteriorBVARPANEL} -> \code{bpvars::PosteriorBVARGROUPPANEL} -> \code{PosteriorBVARGROUPPRIORPANEL}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{last_draw}}{an object of class BVARGROUPPRIORPANEL with the last draw of the 
current MCMC run as the starting value to be passed to the continuation 
of the MCMC estimation using \code{estimate()}.}

\item{\code{posterior}}{a list containing Bayesian estimation output.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PosteriorBVARGROUPPRIORPANEL-new}{\code{specify_posterior_bvarGroupPriorPANEL$new()}}
\item \href{#method-PosteriorBVARGROUPPRIORPANEL-clone}{\code{specify_posterior_bvarGroupPriorPANEL$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bpvars" data-topic="PosteriorBVARPANEL" data-id="get_last_draw"><a href='../../bpvars/html/PosteriorBVARPANEL.html#method-PosteriorBVARPANEL-get_last_draw'><code>bpvars::PosteriorBVARPANEL$get_last_draw()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bpvars" data-topic="PosteriorBVARPANEL" data-id="get_posterior"><a href='../../bpvars/html/PosteriorBVARPANEL.html#method-PosteriorBVARPANEL-get_posterior'><code>bpvars::PosteriorBVARPANEL$get_posterior()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PosteriorBVARGROUPPRIORPANEL-new"></a>}}
\if{latex}{\out{\hypertarget{method-PosteriorBVARGROUPPRIORPANEL-new}{}}}
\subsection{Method \code{new()}}{
Create a new posterior output PosteriorBVARGROUPPRIORPANEL.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_posterior_bvarGroupPriorPANEL$new(specification, posterior)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{specification}}{an object of class BVARGROUPPRIORPANEL with the last 
draw of the current MCMC run as the starting value.}

\item{\code{posterior}}{a list containing Bayesian estimation output.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A posterior output PosteriorBVARGROUPPRIORPANEL.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PosteriorBVARGROUPPRIORPANEL-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PosteriorBVARGROUPPRIORPANEL-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_posterior_bvarGroupPriorPANEL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
