% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.PosteriorBVARGROUPPRIORPANEL}
\alias{summary.PosteriorBVARGROUPPRIORPANEL}
\title{Provides posterior estimation summary for Bayesian Hierarchical Panel 
Vector Autoregressions with group-specific global prior}
\usage{
\method{summary}{PosteriorBVARGROUPPRIORPANEL}(object, ...)
}
\arguments{
\item{object}{an object of class \code{PosteriorBVARGROUPPRIORPANEL} obtained using the
\code{estimate()} function applied to  
Vector Autoregressions containing draws from the  posterior distribution of 
the parameters.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
A list reporting the posterior mean, standard deviations, as well as 5 and 95 
percentiles of the country-specific parameters.
}
\description{
Provides posterior mean, standard deviations, as well as 5 and 95 
percentiles of the parameters for all \code{C} countries.
}
\examples{
# specify the model
specification = specify_bvarGroupPriorPANEL$new(
      data = ilo_dynamic_panel,
      group_allocation = country_grouping_region
)
burn_in       = estimate(specification, 5)             # run the burn-in
posterior     = estimate(burn_in, 5)                   # estimate the model
summary(posterior)

# workflow with the pipe |>
############################################################
ilo_dynamic_panel |>
  specify_bvarGroupPriorPANEL$new(
        exogenous = ilo_exogenous_variables, 
        group_allocation = country_grouping_region) |>
  estimate(S = 5) |> 
  estimate(S = 5) |> 
  summary()

}
\seealso{
\code{\link{estimate.BVARGROUPPRIORPANEL}}, \code{\link{specify_bvarGroupPriorPANEL}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
