% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.PosteriorBVARPANEL}
\alias{summary.PosteriorBVARPANEL}
\title{Provides posterior estimation summary for Bayesian Hierarchical Panel 
Vector Autoregressions}
\usage{
\method{summary}{PosteriorBVARPANEL}(object, ...)
}
\arguments{
\item{object}{an object of class \code{PosteriorBVARPANEL} obtained using the
\code{estimate()} function applied to  
Vector Autoregressions containing draws from the  posterior distribution of 
the parameters.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
A list reporting the posterior mean, standard deviations, as well as 5 and 95 
percentiles of the country-specific parameters.
}
\description{
Provides posterior mean, standard deviations, as well as 5 and 95 
percentiles of the parameters for all \code{C} countries.
}
\examples{
# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel, 
      exogenous = ilo_exogenous_variables)
burn_in       = estimate(specification, 5)             # run the burn-in
posterior     = estimate(burn_in, 5)                   # estimate the model
summary(posterior)

}
\seealso{
\code{\link{estimate.BVARPANEL}}, \code{\link{specify_bvarPANEL}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
